kfence: Refactor reporting into report.c
This is very much WIP, because we now definitely have too many externs.
This will change in subsequent commits, but thought I'd get this out
there so we avoid massive conflicts due to concurrent changes.
