+++ mm/kfence/Makefile
@@ -1,4 +1,5 @@
 # SPDX-License-Identifier: GPL-2.0
-obj-$(CONFIG_KFENCE) := core.o
+
+obj-$(CONFIG_KFENCE) := core.o report.o
 obj-$(CONFIG_KFENCE_STEAL) += steal.o
 obj-$(CONFIG_KFENCE_NAIVE) += naive.o
