+++ mm/kfence/kfence.h
@@ -3,7 +3,9 @@
 #ifndef MM_KFENCE_KFENCE_H
 #define MM_KFENCE_KFENCE_H
 
+#include <linux/mm.h>
 #include <linux/slub_def.h>
+#include <linux/stackdepot.h>
 #include <linux/types.h>
 
 /*
@@ -18,8 +20,33 @@
 		__cond;                                                        \
 	})
 
+enum kfence_object_state {
+	KFENCE_OBJECT_UNUSED,
+	KFENCE_OBJECT_ALLOCATED,
+	KFENCE_OBJECT_FREED
+};
+
+struct kfence_alloc_metadata {
+	depot_stack_handle_t alloc_stack, free_stack;
+	struct kmem_cache *cache;
+	/*
+	 * Size may be read without a lock in ksize(). We assume that ksize() is
+	 * only called for valid (allocated) pointers.
+	 * size>0 means left alignment, size<0 - right alignment.
+	 */
+	int size;
+	/*
+	 * Actual object address. Cannot be calculated from size, because of
+	 * alignment requirements.
+	 */
+	unsigned long addr;
+	enum kfence_object_state state;
+};
+
 extern bool kfence_enabled;
 extern unsigned long kfence_sample_rate;
+extern struct kfence_alloc_metadata *kfence_metadata;
+extern char kfence_dump_buf[PAGE_SIZE * 2];
 
 static inline bool kfence_is_enabled(void)
 {
@@ -32,8 +59,6 @@ void *kfence_guarded_alloc(struct kmem_cache *cache, size_t override_size,
 			   gfp_t gfp);
 void kfence_guarded_free(void *addr);
 
-struct alloc_metadata;
-
 enum kfence_error_kind {
 	KFENCE_ERROR_OOB,
 	KFENCE_ERROR_UAF,
@@ -41,8 +66,11 @@ enum kfence_error_kind {
 };
 
 void kfence_report_error(unsigned long address, int obj_index,
-			 struct alloc_metadata *object,
+			 struct kfence_alloc_metadata *metadata,
 			 enum kfence_error_kind kind);
+
+int kfence_dump_object(char *buf, size_t buf_size, int obj_index,
+		       struct kfence_alloc_metadata *obj);
 
 /* Should be provided by the sampling algorithm implementation. */
 void kfence_impl_init(void);
